Scrolling Test and Collision Grid Changes
Scrolling took a little longer than expected to get it working right in
terms of collision because I realized that I set the collision grid
based on the view instead of the window. After tweaking it with
mapCoordsToPixels and other stuff, I got the bounds to work correctly.

However, when it comes to the scrolling speed itself, all the objects
affected by gameView are moved, including the player, which I did not
expect. I even tried to move the player object based on the scrolling
speed, but the player object still appears to be moving with the
scrolling if left untouched. For this release, I commented the scroll
speed out.

Also, the game runs faster to full 60FPS until the bullets or other
entities collide with Explosion, which reduced the frame rate to around
30FPS again.