Code Fixes
I discovered a simple but big mistake: I put the view scrolling inside
the for loop, which is a big NO in game programming. Scrolling a
layer/view should only happen once per frame, which I did. Also, I fixed
the collision boxes since I mistook width and height of the Rect as
bottom and right corners, which resulted in more collision checks with
improper behavior.

Now I can move on to making the powerups